requirements.txt
flask
yfinance
python-dotenv
requests
python-telegram-bot